adding a test, cleaning other tests in jsonio_atoms

+++ ase/constraints.py
@@ -105,7 +105,7 @@ class IndexedConstraint(FixConstraint):
         if indices.dtype == bool:
             indices = np.arange(len(indices))[indices]
         elif len(indices) == 0:
-            indices = np.empty(0, dtype=np.integer)
+            indices = np.empty(0, dtype=int)
         elif not np.issubdtype(indices.dtype, np.integer):
             raise ValueError('Indices must be integers or boolean mask, '
                              f'not dtype={indices.dtype}')
